move all redux stuf to store folder
