+++ src/store/constants/tags.js
@@ -1,0 +1,3 @@
+export const FILTER_TRACKS_BY_TAGS = "FILTER_TRACKS_BY_TAG";
+export const DELETE_TAG = "DELETE_TAG";
+export const DELETE_ALL_TAGS = "DELETE_ALL_TAGS";
