+++ src/store/constants/tracks.js
@@ -1,0 +1,3 @@
+export const REQUEST_TRACKS = "REQUEST_TRACKS";
+export const GET_TOP_TRACKS_SUCCESS = "GET_TOP_TRACKS_SUCCESS";
+export const GET_TOP_TRACKS_ERROR = "GET_TOP_TRACKS_ERROR";
