+++ src/store/index.js
@@ -2,7 +2,7 @@ import {createStore, applyMiddleware} from "redux";
 import thunk from "redux-thunk";
 import logger from "redux-logger";
 
-import reducer from "../reducers";
+import reducer from "reducers";
 import {loadState, saveState} from "./localStorage";
 
 const initialState = {
